Chore(test): Broke module into two
